Added Text, Form, and running frisbee.
I added the text, fixed the last slide, added a registration form and
started adding a little disc that will be in navbar (fairly early not
working yet).

With just a bit of styling I think it would look really good. Right now
it doesn't look great but it has the programmign behind it.